Refactor event info types
