+++ rmw_gurumdds_cpp/include/rmw_gurumdds_cpp/rmw_context_impl.hpp
@@ -41,6 +41,24 @@
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
 
+namespace rmw_gurumdds_cpp
+{
+void on_participant_changed(
+  const dds_DomainParticipant * a_participant,
+  const dds_ParticipantBuiltinTopicData * data,
+  dds_InstanceHandle_t handle);
+
+void on_publication_changed(
+  const dds_DomainParticipant * a_participant,
+  const dds_PublicationBuiltinTopicData * data,
+  dds_InstanceHandle_t handle);
+
+void on_subscription_changed(
+  const dds_DomainParticipant * a_participant,
+  const dds_SubscriptionBuiltinTopicData * data,
+  dds_InstanceHandle_t handle);
+} // namespace rmw_gurumdds_cpp
+
 struct rmw_context_impl_s
 {
   rmw_dds_common::Context common_ctx;
