+++ rmw_gurumdds_cpp/include/rmw_gurumdds_cpp/types.hpp
@@ -40,28 +40,6 @@
 
 namespace rmw_gurumdds_cpp
 {
-void on_participant_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_ParticipantBuiltinTopicData * data,
-  dds_InstanceHandle_t handle);
-
-void on_publication_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_PublicationBuiltinTopicData * data,
-  dds_InstanceHandle_t handle);
-
-void on_subscription_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_SubscriptionBuiltinTopicData * data,
-  dds_InstanceHandle_t handle);
-
-struct WaitSetInfo
-{
-  dds_WaitSet * wait_set;
-  dds_ConditionSeq * active_conditions;
-  dds_ConditionSeq * attached_conditions;
-};
-
 struct event_callback_data_t
 {
   std::mutex mutex;
