+++ rmw_gurumdds_cpp/src/types.cpp
@@ -14,8 +14,6 @@
 
 #include <cstdint>
 
-#include "rmw/impl/cpp/key_value.hpp"
-
 #include "rmw_gurumdds_cpp/event_converter.hpp"
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/qos.hpp"
@@ -299,188 +297,6 @@ bool PublisherInfo::has_callback_unsafe(rmw_event_type_t event_type) const
 {
   // RMW_EVENT_PUBLISHER_INCOMPATIBLE_TYPE is always used with a callback
   return ((mask | dds_INCONSISTENT_TOPIC_STATUS) & rmw_gurumdds_cpp::get_status_kind_from_rmw(event_type)) > 0;
-}
-
-static inline std::map<std::string, std::vector<uint8_t>>
-parse_map(uint8_t * const data, const uint32_t data_len)
-{
-  std::vector<uint8_t> data_vec(data, data + data_len);
-  std::map<std::string, std::vector<uint8_t>> map
-    = rmw::impl::cpp::parse_key_value(data_vec);
-  return map;
-}
-
-static inline rmw_ret_t
-get_user_data_key(
-  dds_ParticipantBuiltinTopicData * data,
-  const std::string& key,
-  std::string & value,
-  bool & found)
-{
-  found = false;
-  uint8_t * user_data =
-    static_cast<uint8_t *>(data->user_data.value);
-  const uint32_t user_data_len = data->user_data.size;
-  if (user_data_len == 0) {
-    return RMW_RET_OK;
-  }
-
-  auto map = parse_map(user_data, user_data_len);
-  auto name_found = map.find(key);
-  if (name_found != map.end()) {
-    value = std::string(name_found->second.begin(), name_found->second.end());
-    found = true;
-  }
-
-  return RMW_RET_OK;
-}
-
-void on_participant_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_ParticipantBuiltinTopicData * data,
-  dds_InstanceHandle_t handle)
-{
-  auto * participant = const_cast<dds_DomainParticipant *>(a_participant);
-  auto * ctx = reinterpret_cast<rmw_context_impl_t *>(
-    dds_Entity_get_context(reinterpret_cast<dds_Entity *>(participant), 0)
-  );
-  if (ctx == nullptr) {
-    return;
-  }
-
-  rmw_gurumdds_cpp::Guid_t dp_guid{*data};
-  if (handle == dds_HANDLE_NIL) {
-    rmw_gurumdds_cpp::graph_cache::remove_participant(ctx, &dp_guid);
-  } else {
-    std::string enclave_str;
-    bool enclave_found;
-    dds_ReturnCode_t rc =
-      get_user_data_key(
-      const_cast<dds_ParticipantBuiltinTopicData *>(data),
-      "securitycontext", enclave_str, enclave_found);
-    if (RMW_RET_OK != rc) {
-      RMW_SET_ERROR_MSG("failed to parse user data for enclave");
-    }
-
-    const char * enclave = nullptr;
-    if (enclave_found) {
-      enclave = enclave_str.c_str();
-    }
-
-    if (RMW_RET_OK != rmw_gurumdds_cpp::graph_cache::add_participant(ctx, &dp_guid, enclave)) {
-      RMW_SET_ERROR_MSG("failed to assert remote participant in graph");
-    }
-  }
-}
-
-void on_publication_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_PublicationBuiltinTopicData * data,
-  dds_InstanceHandle_t handle)
-{
-  auto * participant = const_cast<dds_DomainParticipant *>(a_participant);
-  auto * ctx = reinterpret_cast<rmw_context_impl_t *>(
-    dds_Entity_get_context(reinterpret_cast<dds_Entity *>(participant), 0)
-  );
-  if (ctx == nullptr) {
-    return;
-  }
-
-  rmw_gurumdds_cpp::Guid_t dp_guid = rmw_gurumdds_cpp::Guid_t::for_participant(*data);
-  rmw_gurumdds_cpp::Guid_t endp_guid{*data};
-  const auto * dp_guid_prefix = reinterpret_cast<const uint32_t *>(dp_guid.prefix);
-  const auto * endp_guid_prefix = reinterpret_cast<const uint32_t *>(endp_guid.prefix);
-  if (handle == dds_HANDLE_NIL) {
-    RCUTILS_LOG_DEBUG_NAMED(
-      "pub on data available",
-      "[ud] endp_gid=0x%08X.0x%08X.0x%08X.0x%08X ",
-      endp_guid_prefix[0],
-      endp_guid_prefix[1],
-      endp_guid_prefix[2],
-      endp_guid.entityId);
-    rmw_gurumdds_cpp::graph_cache::remove_entity(ctx, &endp_guid, false);
-  } else {
-    rmw_gurumdds_cpp::graph_cache::add_remote_entity(
-      ctx,
-      &endp_guid,
-      &dp_guid,
-      data->topic_name,
-      data->type_name,
-      data->user_data,
-      &data->reliability,
-      &data->durability,
-      &data->deadline,
-      &data->liveliness,
-      &data->lifespan,
-      false);
-    RCUTILS_LOG_DEBUG_NAMED(
-      "pub on data available",
-      "dp_gid=0x%08X.0x%08X.0x%08X.0x%08X, "
-      "gid=0x%08X.0x%08X.0x%08X.0x%08X, ",
-      dp_guid_prefix[0],
-      dp_guid_prefix[1],
-      dp_guid_prefix[2],
-      dp_guid.entityId,
-      endp_guid_prefix[0],
-      endp_guid_prefix[1],
-      endp_guid_prefix[2],
-      endp_guid.entityId);
-  }
-}
-
-void on_subscription_changed(
-  const dds_DomainParticipant * a_participant,
-  const dds_SubscriptionBuiltinTopicData * data,
-  dds_InstanceHandle_t handle)
-{
-  auto * participant = const_cast<dds_DomainParticipant *>(a_participant);
-  auto * ctx = reinterpret_cast<rmw_context_impl_t *>(
-    dds_Entity_get_context(reinterpret_cast<dds_Entity *>(participant), 0)
-  );
-  if (ctx == nullptr) {
-    return;
-  }
-
-  rmw_gurumdds_cpp::Guid_t dp_guid = rmw_gurumdds_cpp::Guid_t::for_participant(*data);
-  rmw_gurumdds_cpp::Guid_t endp_guid{*data};
-  const auto * dp_guid_prefix = reinterpret_cast<const uint32_t *>(dp_guid.prefix);
-  const auto * endp_guid_prefix = reinterpret_cast<const uint32_t *>(endp_guid.prefix);
-  if (handle == dds_HANDLE_NIL) {
-    RCUTILS_LOG_DEBUG_NAMED(
-      "sub on data available",
-      "[ud] endp_gid=0x%08X.0x%08X.0x%08X.0x%08X ",
-      endp_guid_prefix[0],
-      endp_guid_prefix[1],
-      endp_guid_prefix[2],
-      endp_guid.entityId);
-    rmw_gurumdds_cpp::graph_cache::remove_entity(ctx, &endp_guid, false);
-  } else {
-    rmw_gurumdds_cpp::graph_cache::add_remote_entity(
-      ctx,
-      &endp_guid,
-      &dp_guid,
-      data->topic_name,
-      data->type_name,
-      data->user_data,
-      &data->reliability,
-      &data->durability,
-      &data->deadline,
-      &data->liveliness,
-      nullptr,
-      true);
-    RCUTILS_LOG_DEBUG_NAMED(
-      "sub on data available",
-      "dp_gid=0x%08X.0x%08X.0x%08X.0x%08X, "
-      "gid=0x%08X.0x%08X.0x%08X.0x%08X, ",
-      dp_guid_prefix[0],
-      dp_guid_prefix[1],
-      dp_guid_prefix[2],
-      dp_guid.entityId,
-      endp_guid_prefix[0],
-      endp_guid_prefix[1],
-      endp_guid_prefix[2],
-      endp_guid.entityId);
-  }
 }
 
 rmw_ret_t SubscriberInfo::set_on_new_event_callback(
